Change base path
As the resulting site is hosted as a github page, it might be easier to
manage to make the base path points to the github page and let the DNS
settings do the custom domain stuff.

This should fix the JavaScript and CSS not loading and internal link
directs to nomacs.org issue.

Fixes https://github.com/nomacs/nomacs.github.io/issues/6.